+++ bagel/mappings.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 from pathlib import Path
 
-from bagel.utility import load_json
+import bagel.file_utils as futil
 
 Namespace = namedtuple("Namespace", ["pf", "url"])
 COGATLAS = Namespace("cogatlas", "https://www.cognitiveatlas.org/task/id/")
@@ -48,7 +48,7 @@ def get_pipeline_uris() -> dict:
     """
     output_dict = {}
     for pipe_file in PROCESSING_PIPELINE_PATH.glob("*.json"):
-        pipe = load_json(pipe_file)
+        pipe = futil.load_json(pipe_file)
         output_dict[pipe["name"]] = f"{NP.pf}: {pipe['name']}"
 
     return output_dict
@@ -61,7 +61,7 @@ def get_pipeline_versions() -> dict:
     """
     output_dict = {}
     for pipe_file in PROCESSING_PIPELINE_PATH.glob("*.json"):
-        pipe = load_json(pipe_file)
+        pipe = futil.load_json(pipe_file)
         output_dict[pipe["name"]] = pipe["versions"]
 
     return output_dict
